Remove background_jobs_enabled, move code from tenant_mgr.rs to tenant.rs
